Commit: Create Auction Flow. Document (please read): https://www.notion.so/L-p-tr-nh-n-ng-cao-344603902108803eaf23caddc4a34187#34b60390210880698c07da06d8a8e2a4
- Added ProductImageDTO class with a new line for formatting.
- Created ProductImageMapper for mapping ProductImageDTO to ProductImage.
- Introduced ProductMapper as a placeholder for future product mapping logic.
- Implemented AuctionRowMapper for mapping ResultSet to Auction model.
- Developed ProductImageRowMapper for mapping ResultSet to ProductImage model.
- Created ProductRowMapper for mapping ResultSet to Product model, with a note on image handling.
- Added UserRowMapper for mapping ResultSet to User model.
- Implemented JdbcAuctionRepository for auction data persistence with methods for saving and retrieving auctions.
- Developed JdbcProductImageRepository for managing product images in the database.
- Created JdbcProductRepository for product data persistence with methods for saving and retrieving products.
- Implemented JdbcUserRepository for user data persistence with methods for finding and saving users.
- Added ProductImageRepository interface for product image operations.
- Created ProductRepository interface for product operations.
- Added AuctionServiceIntegrationTest for testing auction service functionality.
- Implemented AuctionServiceTest for unit testing auction service validation logic.
- Introduced PublishAuctionsRequest DTO for auction publishing requests.

+++ server/src/main/java/com/vbay/server/Model/Bid.java
@@ -1,17 +1,19 @@
 package com.vbay.server.Model;
 
+import java.math.BigDecimal;
+
 public class Bid {
     // Base
     private long id;
     private long auction_id;
     private long bidder_id;
     // Bid Detail
-    private double bid_amount;
+    private BigDecimal bid_amount;
     private String bid_time;
     private boolean is_winner;
 
     //Constructor (Create new bid)
-    public Bid(long auction_id, long bidder_id, double bid_amount, String bid_time){
+    public Bid(long auction_id, long bidder_id, BigDecimal bid_amount, String bid_time){
         //Base:
         this.auction_id = auction_id;
         this.bidder_id = bidder_id;
@@ -33,7 +35,7 @@ public class Bid {
         return this.bidder_id;
     }
     //b, Bid Detail:
-    public double getBidAmount(){
+    public BigDecimal getBidAmount(){
         return this.bid_amount;
     }
     public String getBidTime(){
@@ -55,7 +57,7 @@ public class Bid {
         this.bidder_id = new_bidder_id;
     }
     //b, Bid Detail:
-    public void setBidAmount(double new_bid_amount){
+    public void setBidAmount(BigDecimal new_bid_amount){
         this.bid_amount = new_bid_amount;
     }
     public void setBidTime(String new_bid_time){
